make it possible to change options (except for resolution); create folder for set of options

+++ src/ImageCreator.java
@@ -32,39 +32,42 @@ public class ImageCreator {
 		int imgCount = 1;
 		
 		for (int i = 0; i < liste.length; i++){
-			Mat tmp1 = Imgcodecs.imread(liste[i].getAbsolutePath());
-			// rotate image?
-			if (tmp1.rows() > tmp1.cols()) {
-				tmp1 = rotate90(tmp1);
-			}
-			// resize
-			float format = (float) tmp1.cols() / (float) tmp1.rows();
-
-			if (format == collageOptions.getFormat()) {
-				Imgproc.resize(tmp1, tmp1, new Size(collageOptions.getPartWidth(), collageOptions.getPartHeight()));
-			} else if (format < collageOptions.getFormat()) {
-				float newWidth = tmp1.cols() * ((float) collageOptions.getPartHeight() / tmp1.rows());
-				Imgproc.resize(tmp1, tmp1, new Size(newWidth, collageOptions.getPartHeight()));
-			} else {
-				float newHeight = tmp1.rows() * ((float) collageOptions.getPartWidth() / tmp1.cols());
-				Imgproc.resize(tmp1, tmp1, new Size(collageOptions.getPartWidth(), newHeight));
-			}
-			
-			// copy
-			Mat destinationROI = new Mat(result, new Rect (new Point(x, y), tmp1.size()));
-			tmp1.copyTo(destinationROI);
-			x += collageOptions.getPartWidth();
-			if (x >= collageOptions.getResultWidth()) {
-				x = 0;
-				y += collageOptions.getPartHeight();
-			}
-			if (y >= collageOptions.getResultHeight() || i == liste.length - 1) {
-				// save image
-				Imgcodecs.imwrite(dir.getAbsolutePath() + collageOptions.getFileName(imgCount), result);
-				result = new Mat(collageOptions.getResultHeight(), collageOptions.getResultWidth(), CvType.CV_8UC3, new Scalar(0, 0, 0));
-				x = 0;
-				y = 0;
-				imgCount++;
+			if (liste[i].isFile()) {
+				Mat tmp1 = Imgcodecs.imread(liste[i].getAbsolutePath());
+				// rotate image?
+				if (tmp1.rows() > tmp1.cols()) {
+					tmp1 = rotate90(tmp1);
+				}
+				// resize
+				float format = (float) tmp1.cols() / (float) tmp1.rows();
+	
+				if (format == collageOptions.getFormat()) {
+					Imgproc.resize(tmp1, tmp1, new Size(collageOptions.getPartWidth(), collageOptions.getPartHeight()));
+				} else if (format < collageOptions.getFormat()) {
+					float newWidth = tmp1.cols() * ((float) collageOptions.getPartHeight() / tmp1.rows());
+					Imgproc.resize(tmp1, tmp1, new Size(newWidth, collageOptions.getPartHeight()));
+				} else {
+					float newHeight = tmp1.rows() * ((float) collageOptions.getPartWidth() / tmp1.cols());
+					Imgproc.resize(tmp1, tmp1, new Size(collageOptions.getPartWidth(), newHeight));
+				}
+				
+				// copy
+				Mat destinationROI = new Mat(result, new Rect (new Point(x, y), tmp1.size()));
+				tmp1.copyTo(destinationROI);
+				x += collageOptions.getPartWidth();
+				if (x >= collageOptions.getResultWidth()) {
+					x = 0;
+					y += collageOptions.getPartHeight();
+				}
+				if (y >= collageOptions.getResultHeight() || i == liste.length - 1) {
+					// save image
+					Imgcodecs.imwrite(dir.getAbsolutePath() + collageOptions.getFileName(imgCount), result);
+					System.out.println(dir.getAbsolutePath() + collageOptions.getFileName(imgCount));
+					result = new Mat(collageOptions.getResultHeight(), collageOptions.getResultWidth(), CvType.CV_8UC3, new Scalar(0, 0, 0));
+					x = 0;
+					y = 0;
+					imgCount++;
+				}
 			}
 		}
 //		showImage(mat2Img(result));
